refactor: extract API routes and simplify RAG pipeline
Split monolithic main.py into route modules (health, laws, legislation,
query, threads). Add dedicated query_service for RAG orchestration.
Rewrite citation_filter to use regex-based reference extraction instead
of LLM calls. Simplify qdrant_service by removing unused methods.
Add batch message count query to conversation_service.
