+++ app/database.py
@@ -1,4 +1,5 @@
 import os
+from collections.abc import Generator
 
 from sqlmodel import Session, SQLModel, create_engine
 
@@ -9,9 +10,11 @@ engine = create_engine(DATABASE_URL, connect_args=connect_args)
 
 
 def create_db_and_tables() -> None:
+    """Create all SQLModel tables if they don't already exist."""
     SQLModel.metadata.create_all(engine)
 
 
-def get_session():
+def get_session() -> Generator[Session, None, None]:
+    """FastAPI dependency that yields a SQLModel session per request."""
     with Session(engine) as session:
         yield session
